+++ packages/components/src/components/post-header/post-header.tsx
@@ -42,7 +42,7 @@ export class PostHeader {
   private localHeaderResizeObserver: ResizeObserver;
   get scrollParent(): HTMLElement {
     const frozenScrollParent: HTMLElement | null = document.querySelector(
-      '[data-is-post-header-scroll-parent]',
+      '[data-post-scroll-locked]',
     );
 
     if (frozenScrollParent) return frozenScrollParent;
@@ -59,7 +59,7 @@ export class PostHeader {
       element = element.parentElement;
     }
 
-    return document.documentElement;
+    return document.body;
   }
 
   @Element() host: HTMLPostHeaderElement;
@@ -69,17 +69,18 @@ export class PostHeader {
 
   @State() megadropdownOpen: boolean = false;
 
+  @Watch('device')
   @Watch('mobileMenuExtended')
-  frozeBody(isMobileMenuExtended: boolean) {
+  lockBody(newValue: boolean | string, _oldValue: boolean | string, propName: string) {
     const scrollParent = this.scrollParent;
+    const mobileMenuExtended =
+      propName === 'mobileMenuExtended' ? newValue : this.mobileMenuExtended;
 
-    if (isMobileMenuExtended) {
-      scrollParent.setAttribute('data-is-post-header-scroll-parent', '');
-      scrollParent.style.overflow = 'hidden';
+    if (this.device !== 'desktop' && mobileMenuExtended) {
+      scrollParent.setAttribute('data-post-scroll-locked', '');
       this.host.addEventListener('keydown', this.keyboardHandler);
     } else {
-      scrollParent.style.overflow = '';
-      scrollParent.removeAttribute('data-is-post-header-scroll-parent');
+      scrollParent.removeAttribute('data-post-scroll-locked');
       this.host.removeEventListener('keydown', this.keyboardHandler);
     }
   }
@@ -109,19 +110,15 @@ export class PostHeader {
     document.addEventListener('postToggleMegadropdown', this.megedropdownStateHandler);
     this.host.addEventListener('click', this.handleLinkClick);
 
-    this.frozeBody(false);
     this.handleResize();
     this.handleScrollEvent();
     this.handleScrollParentResize();
+    this.lockBody(false, this.mobileMenuExtended, 'mobileMenuExtended');
   }
 
   componentDidRender() {
     this.getFocusableElements();
     this.handleLocalHeaderResize();
-  }
-
-  componentDidLoad() {
-    // Check if the mega dropdown is expanded
   }
 
   // Clean up possible side effects when post-header is disconnected
@@ -165,7 +162,12 @@ export class PostHeader {
     // Toggle menu visibility before it slides down and after it slides back up
     if (this.mobileMenuExtended) await this.mobileMenuAnimation.finished;
     this.mobileMenuExtended = force ?? !this.mobileMenuExtended;
-    if (!this.mobileMenuExtended) await this.mobileMenuAnimation.finished;
+
+    if (this.mobileMenuExtended === false) {
+      Array.from(this.host.querySelectorAll('post-megadropdown')).forEach(dropdown => {
+        dropdown.hide(false, true);
+      });
+    }
   }
 
   private megedropdownStateHandler(event: CustomEvent) {
@@ -214,7 +216,9 @@ export class PostHeader {
   }
 
   private handleScrollEvent() {
-    this.host.style.setProperty('--header-scroll-top', `${this.scrollParent.scrollTop}px`);
+    const scrollTop =
+      this.scrollParent === document.body ? window.scrollY : this.scrollParent.scrollTop;
+    this.host.style.setProperty('--header-scroll-top', `${scrollTop}px`);
   }
 
   private updateLocalHeaderHeight() {
@@ -265,12 +269,6 @@ export class PostHeader {
       newDevice = 'mobile';
     }
 
-    // Close any open mobile menu
-    if (newDevice === 'desktop' && this.mobileMenuExtended) {
-      this.toggleMobileMenu();
-      this.mobileMenuAnimation.finish(); // no animation
-    }
-
     // Apply only on change for doing work only when necessary
     if (newDevice !== previousDevice) {
       this.device = newDevice;
@@ -307,15 +305,22 @@ export class PostHeader {
 
   private renderNavigation() {
     const navigationClasses = ['navigation'];
+
+    const mobileMenuScrollTop = this.mobileMenu?.scrollTop ?? 0;
+
     if (this.mobileMenuExtended) {
       navigationClasses.push('extended');
     }
-    if (!this.megadropdownOpen) {
-      navigationClasses.push('scroll-y');
+
+    if (this.megadropdownOpen) {
+      navigationClasses.push('megadropdown-open');
     }
 
     return (
-      <div class={navigationClasses.join(' ')}>
+      <div
+        class={navigationClasses.join(' ')}
+        style={{ '--header-navigation-current-inset': `${mobileMenuScrollTop}px` }}
+      >
         <div class="mobile-menu" ref={el => (this.mobileMenu = el)}>
           <slot name="post-mainnavigation"></slot>