+++ packages/components/src/components/post-list-item/post-list-item.tsx
@@ -1,4 +1,5 @@
 import { Component, Element, Host, h } from '@stencil/core';
+import { version } from '@root/package.json';
 
 /**
  * @slot default- Slot for placing the content of the list item.
@@ -18,7 +19,7 @@ export class PostListItem {
 
   render() {
     return (
-      <Host role="listitem" slot="post-list-item">
+      <Host role="listitem" slot="post-list-item" version={version}>
         <slot></slot>
       </Host>
     );