+++ packages/icons/src/utilities/environment.ts
@@ -3,6 +3,7 @@
 import * as dotenv from 'dotenv';
 import { HttpsProxyAgent } from 'https-proxy-agent';
 import { RequestInit } from 'node-fetch';
+import { MESSAGE_ENV_VARS_MISSING_ERROR } from './constants';
 
 dotenv.config();
 const user = process.env.CEN_USERNAME;
@@ -10,24 +11,22 @@ const pw = process.env.CEN_PASSWORD;
 const proxy = process.env.HTTPS_PROXY;
 const passphrase = Buffer.from(`${user}:${pw}`).toString('base64');
 
-export const url = process.env.CEN_URL;
+export const urls = {
+  post: process.env.CEN_URL_ICONSET_POST ?? '',
+  ui: process.env.CEN_URL_ICONSET_UI ?? '',
+};
 
-if (!user || !pw || !url) {
-  throw new Error(
-    'Environment variables are not defined. Please check your .env file and compare it to the .template.env. Are there any variables missing or undefined?',
-  );
-}
+export const requestInit: RequestInit = {
+  headers: {
+    Authorization: `Basic ${passphrase}`,
+  },
+  agent: proxy ? new HttpsProxyAgent(proxy) : undefined,
+};
 
-export const getRequestInit = () => {
-  const request: RequestInit = {
-    headers: {
-      Authorization: `Basic ${passphrase}`,
-    },
-  };
-
-  if (proxy) {
-    request.agent = new HttpsProxyAgent(proxy);
+export function checkEnvVarsExist(): boolean {
+  if (!user || !pw || Object.values(urls).some(v => !v)) {
+    throw new Error(MESSAGE_ENV_VARS_MISSING_ERROR);
   }
 
-  return request;
-};
+  return true;
+}